+++ app/src/main/java/com/danosoftware/galaxyforce/sprites/game/aliens/AbstractAlien.java
@@ -1,5 +1,10 @@
 package com.danosoftware.galaxyforce.sprites.game.aliens;
 
+import static com.danosoftware.galaxyforce.sprites.game.aliens.enums.AlienState.ACTIVE;
+import static com.danosoftware.galaxyforce.sprites.game.aliens.enums.AlienState.DESTROYED;
+import static com.danosoftware.galaxyforce.sprites.game.aliens.enums.AlienState.EXPLODING;
+import static com.danosoftware.galaxyforce.sprites.game.aliens.enums.AlienState.WAITING;
+
 import com.danosoftware.galaxyforce.sprites.common.AbstractCollidingSprite;
 import com.danosoftware.galaxyforce.sprites.game.aliens.enums.AlienState;
 import com.danosoftware.galaxyforce.sprites.game.behaviours.explode.ExplodeBehaviour;
@@ -11,11 +16,6 @@ import com.danosoftware.galaxyforce.sprites.game.behaviours.spinner.SpinningBeha
 import com.danosoftware.galaxyforce.sprites.game.missiles.bases.IBaseMissile;
 import com.danosoftware.galaxyforce.view.Animation;
 import com.danosoftware.galaxyforce.waves.AlienCharacter;
-
-import static com.danosoftware.galaxyforce.sprites.game.aliens.enums.AlienState.ACTIVE;
-import static com.danosoftware.galaxyforce.sprites.game.aliens.enums.AlienState.DESTROYED;
-import static com.danosoftware.galaxyforce.sprites.game.aliens.enums.AlienState.EXPLODING;
-import static com.danosoftware.galaxyforce.sprites.game.aliens.enums.AlienState.WAITING;
 
 public abstract class AbstractAlien extends AbstractCollidingSprite implements IAlien {
 
@@ -59,27 +59,27 @@ public abstract class AbstractAlien extends AbstractCollidingSprite implements I
     private final AlienCharacter character;
 
     protected AbstractAlien(
-            AlienCharacter character,
-            Animation animation,
-            int x,
-            int y,
-            int energy,
-            FireBehaviour fireBehaviour,
-            PowerUpBehaviour powerUpBehaviour,
-            SpawnBehaviour spawnBehaviour,
-            HitBehaviour hitBehaviour,
-            ExplodeBehaviour explodeBehaviour,
-            SpinningBehaviour spinningBehaviour) {
+        AlienCharacter character,
+        Animation animation,
+        float x,
+        float y,
+        int energy,
+        FireBehaviour fireBehaviour,
+        PowerUpBehaviour powerUpBehaviour,
+        SpawnBehaviour spawnBehaviour,
+        HitBehaviour hitBehaviour,
+        ExplodeBehaviour explodeBehaviour,
+        SpinningBehaviour spinningBehaviour) {
 
-        super(
-                animation.getKeyFrame(
-                        0,
-                        Animation.ANIMATION_LOOPING),
-                x,
-                y);
-        this.character = character;
-        state = ACTIVE;
-        this.energy = energy;
+      super(
+          animation.getKeyFrame(
+              0,
+              Animation.ANIMATION_LOOPING),
+          x,
+          y);
+      this.character = character;
+      state = ACTIVE;
+      this.energy = energy;
         this.explodeBehaviour = explodeBehaviour;
         this.fireBehaviour = fireBehaviour;
         this.powerUpBehaviour = powerUpBehaviour;
