+++ app/src/main/java/com/danosoftware/galaxyforce/sprites/game/aliens/AbstractResettableAlien.java
@@ -1,5 +1,7 @@
 package com.danosoftware.galaxyforce.sprites.game.aliens;
 
+import static com.danosoftware.galaxyforce.sprites.game.aliens.enums.AlienState.FINISHED_PASS;
+
 import com.danosoftware.galaxyforce.sprites.game.behaviours.explode.ExplodeBehaviour;
 import com.danosoftware.galaxyforce.sprites.game.behaviours.fire.FireBehaviour;
 import com.danosoftware.galaxyforce.sprites.game.behaviours.hit.HitBehaviour;
@@ -9,35 +11,33 @@ import com.danosoftware.galaxyforce.sprites.game.behaviours.spinner.SpinningBeha
 import com.danosoftware.galaxyforce.view.Animation;
 import com.danosoftware.galaxyforce.waves.AlienCharacter;
 
-import static com.danosoftware.galaxyforce.sprites.game.aliens.enums.AlienState.FINISHED_PASS;
-
 /**
  * Abstract alien that can be reset and replayed when alien moves off-screen.
  */
 public abstract class AbstractResettableAlien extends AbstractAlien implements IResettableAlien {
 
     protected AbstractResettableAlien(
-            final AlienCharacter character,
-            final Animation animation,
-            final int x,
-            final int y,
-            final int energy,
-            final FireBehaviour fireBehaviour,
-            final PowerUpBehaviour powerUpBehaviour,
-            final SpawnBehaviour spawnBehaviour,
-            final HitBehaviour hitBehaviour,
-            final ExplodeBehaviour explodeBehaviour,
-            final SpinningBehaviour spinningBehaviour) {
+        final AlienCharacter character,
+        final Animation animation,
+        final float x,
+        final float y,
+        final int energy,
+        final FireBehaviour fireBehaviour,
+        final PowerUpBehaviour powerUpBehaviour,
+        final SpawnBehaviour spawnBehaviour,
+        final HitBehaviour hitBehaviour,
+        final ExplodeBehaviour explodeBehaviour,
+        final SpinningBehaviour spinningBehaviour) {
 
-        super(
-                character,
-                animation,
-                x,
-                y,
-                energy,
-                fireBehaviour,
-                powerUpBehaviour,
-                spawnBehaviour,
+      super(
+          character,
+          animation,
+          x,
+          y,
+          energy,
+          fireBehaviour,
+          powerUpBehaviour,
+          spawnBehaviour,
                 hitBehaviour,
                 explodeBehaviour,
                 spinningBehaviour);
