+++ app/src/main/java/com/danosoftware/galaxyforce/sprites/game/bases/IBasePrimary.java
@@ -3,16 +3,14 @@ package com.danosoftware.galaxyforce.sprites.game.bases;
 import com.danosoftware.galaxyforce.models.screens.background.RgbColour;
 import com.danosoftware.galaxyforce.sprites.game.bases.enums.BaseLean;
 import com.danosoftware.galaxyforce.sprites.game.bases.enums.HelperSide;
-
 import java.util.List;
 
 public interface IBasePrimary extends IBase {
 
     /**
-     * Set the wanted target position of the base.
-     * Base will gradually move to this position.
+     * Set the wanted target position of the base. Base will gradually move to this position.
      */
-    void moveTarget(int targetX, int targetY);
+    void moveTarget(float targetX, float targetY);
 
     void helperExploding(HelperSide side);
