+++ app/src/main/java/com/danosoftware/galaxyforce/sprites/game/missiles/aliens/AlienMissileGuided.java
@@ -1,12 +1,12 @@
 package com.danosoftware.galaxyforce.sprites.game.missiles.aliens;
 
-import com.danosoftware.galaxyforce.enumerations.AlienMissileSpeed;
-import com.danosoftware.galaxyforce.sprites.game.bases.IBasePrimary;
-import com.danosoftware.galaxyforce.view.Animation;
-
 import static com.danosoftware.galaxyforce.sprites.game.missiles.aliens.AlienMissileRotater.calculateAngle;
 import static com.danosoftware.galaxyforce.sprites.game.missiles.aliens.AlienMissileRotater.calculateRotation;
 import static com.danosoftware.galaxyforce.utilities.OffScreenTester.offScreenAnySide;
+
+import com.danosoftware.galaxyforce.enumerations.AlienMissileSpeed;
+import com.danosoftware.galaxyforce.sprites.game.bases.IBasePrimary;
+import com.danosoftware.galaxyforce.view.Animation;
 
 /**
  * Alien missile that targets the supplied base.
@@ -15,48 +15,48 @@ import static com.danosoftware.galaxyforce.utilities.OffScreenTester.offScreenAn
  */
 public class AlienMissileGuided extends AbstractAlienMissile {
 
-    /* maximum missile change direction in radians */
-    private static final float MAX_DIRECTION_CHANGE_ANGLE = 0.3f;
+  /* maximum missile change direction in radians */
+  private static final float MAX_DIRECTION_CHANGE_ANGLE = 0.3f;
 
-    /* time delay between missile direction changes */
-    private static final float MISSILE_DIRECTION_CHANGE_DELAY = 0.1f;
+  /* time delay between missile direction changes */
+  private static final float MISSILE_DIRECTION_CHANGE_DELAY = 0.1f;
 
-    /* reference to base to target.
-    /* (this base would have been active when missile was constructed but
-    /* could later target an old base position if destroyed) */
-    /* shouldn't be a problem but missile will target an old base position in this scenario */
-    private final IBasePrimary base;
+  /* reference to base to target.
+  /* (this base would have been active when missile was constructed but
+  /* could later target an old base position if destroyed) */
+  /* shouldn't be a problem but missile will target an old base position in this scenario */
+  private final IBasePrimary base;
 
-    // offset applied to x and y every move
-    private int xDelta;
-    private int yDelta;
+  // offset applied to x and y every move
+  private float xDelta;
+  private float yDelta;
 
-    /* current angle of missile direction */
-    private float angle;
+  /* current angle of missile direction */
+  private float angle;
 
-    /* missile speed in pixels per seconds */
-    private final int speed;
+  /* missile speed in pixels per seconds */
+  private final int speed;
 
-    /* variable to store time passed since last missile direction change */
-    private float timeSinceMissileDirectionChange;
+  /* variable to store time passed since last missile direction change */
+  private float timeSinceMissileDirectionChange;
 
-    // created rotated missile aimed at base
-    public AlienMissileGuided(
-            int xStart,
-            int yStart,
-            final Animation animation,
-            final AlienMissileSpeed missileSpeed,
-            final IBasePrimary base) {
+  // created rotated missile aimed at base
+  public AlienMissileGuided(
+      float xStart,
+      float yStart,
+      final Animation animation,
+      final AlienMissileSpeed missileSpeed,
+      final IBasePrimary base) {
 
-        super(
-                animation,
-                xStart,
-                yStart);
+    super(
+        animation,
+        xStart,
+        yStart);
 
-        this.base = base;
+    this.base = base;
 
-        // initial angle (i.e. straight down)
-        this.angle = (float) Math.atan2(-1, 0);
+    // initial angle (i.e. straight down)
+    this.angle = (float) Math.atan2(-1, 0);
         this.speed = missileSpeed.getSpeed();
         calculateMovements();
 
@@ -72,21 +72,21 @@ public class AlienMissileGuided extends AbstractAlienMissile {
          * Guide missile every x seconds so the missile changes direction to
          * follow any changes in the base's position.
          */
-        timeSinceMissileDirectionChange += deltaTime;
-        if (timeSinceMissileDirectionChange > MISSILE_DIRECTION_CHANGE_DELAY) {
-            targetBase();
-            timeSinceMissileDirectionChange = 0f;
-        }
+      timeSinceMissileDirectionChange += deltaTime;
+      if (timeSinceMissileDirectionChange > MISSILE_DIRECTION_CHANGE_DELAY) {
+        targetBase();
+        timeSinceMissileDirectionChange = 0f;
+      }
 
-        // move missile by calculated deltas
-        moveByDelta(
-                (int) (xDelta * deltaTime),
-                (int) (yDelta * deltaTime));
+      // move missile by calculated deltas
+      moveByDelta(
+          xDelta * deltaTime,
+          yDelta * deltaTime);
 
-        // if missile is now off screen then destroy it
-        if (offScreenAnySide(this)) {
-            destroy();
-        }
+      // if missile is now off screen then destroy it
+      if (offScreenAnySide(this)) {
+        destroy();
+      }
     }
 
     /**
@@ -119,11 +119,11 @@ public class AlienMissileGuided extends AbstractAlienMissile {
     }
 
     private void calculateMovements() {
-        // calculate sprite rotation for wanted angle
-        rotate((calculateRotation(angle)));
+      // calculate sprite rotation for wanted angle
+      rotate((calculateRotation(angle)));
 
-        // calculate the deltas to be applied each move
-        this.xDelta = (int) (this.speed * (float) Math.cos(angle));
-        this.yDelta = (int) (this.speed * (float) Math.sin(angle));
+      // calculate the deltas to be applied each move
+      this.xDelta = this.speed * (float) Math.cos(angle);
+      this.yDelta = this.speed * (float) Math.sin(angle);
     }
 }
