+++ app/src/main/java/com/danosoftware/galaxyforce/sprites/game/missiles/bases/BaseMissileUpwards.java
@@ -1,9 +1,9 @@
 package com.danosoftware.galaxyforce.sprites.game.missiles.bases;
 
+import static com.danosoftware.galaxyforce.utilities.OffScreenTester.offScreenTop;
+
 import com.danosoftware.galaxyforce.enumerations.BaseMissileSpeed;
 import com.danosoftware.galaxyforce.view.Animation;
-
-import static com.danosoftware.galaxyforce.utilities.OffScreenTester.offScreenTop;
 
 /**
  * Missile that fires straight upwards
@@ -14,21 +14,21 @@ public class BaseMissileUpwards extends AbstractBaseMissile {
     private final int missileSpeed;
 
     public BaseMissileUpwards(
-            final int xStart,
-            final int yStart,
-            final Animation animation,
-            final BaseMissileSpeed baseMissileSpeed) {
-        super(
-                animation,
-                xStart,
-                yStart);
-        this.missileSpeed = baseMissileSpeed.getSpeed();
+        final float xStart,
+        final float yStart,
+        final Animation animation,
+        final BaseMissileSpeed baseMissileSpeed) {
+      super(
+          animation,
+          xStart,
+          yStart);
+      this.missileSpeed = baseMissileSpeed.getSpeed();
     }
 
     @Override
     public void animate(float deltaTime) {
 
-        moveYByDelta((int) (missileSpeed * deltaTime));
+      moveYByDelta(missileSpeed * deltaTime);
 
         // if missile is now off screen then destroy it
         if (offScreenTop(this)) {
